% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{package}
\name{bSantiago-package}
\alias{bSantiago}
\alias{bSantiago-package}
\title{bSantiago: A data package with the results of a travel and well-being survey conducted in Santiago in 2016}
\description{
A dataset containing information about socio-demographics, health-related, perceptions and travel behaviour, travel choices and planning, social interaction factors, built environment of users of active and motorized modes of transportation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/bSantiago}
  \item \url{https://paezha.github.io/bSantiago/}
}

}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Niloofar Nalaee \email{nalaeen@mcmaster.ca} (\href{https://orcid.org/0000-0003-4344-0412}{ORCID})
  \item Beatriz Mella Lira \email{beatriz.mellalira@uc.cl} (\href{https://orcid.org/0000-0002-5248-2684}{ORCID})
}

}
\keyword{internal}
