% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{comunas_1}
\alias{comunas_1}
\title{A data frame with geographic information of respondents to the survey}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 52 rows and 11 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(comunas_1)
}
\description{
A dataset containing geographic information of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 52 rows and 11 variables:
\describe{
\item{OBJECTID}{Unique identifier of respondent}
\item{ID}{Unique identifier of respondent}
\item{COMUNA}{Commune of residence}
\item{REGION}{Region of residence}
\item{Ha}{Height of residence}
\item{LAYER}{Layer of residence}
\item{COORD_X}{X coordinate of the place where respondents live}
\item{COORD_X}{Y coordinate of the place where respondents live}
\item{Shape_Leng}{Length of the ploygons where respondents live}
\item{Shape_Area}{Area of the ploygons where respondents live}
\item{geometry}{Geometry of the ploygons where respondents live}
}
}
\examples{
 data(comunas_1)
 COMUNA <- comunas_1$COMUNA
}
\keyword{datasets}
