% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{load_data_from_rda}
\alias{load_data_from_rda}
\title{Load data from an RDA file.}
\usage{
load_data_from_rda()
}
\value{
A data frame containing the loaded data.
}
\description{
Load data from an RDA file.
}
