% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{San_comuna2}
\alias{San_comuna2}
\title{A data frame with information about whole survey and geographic features of respondents to the survey}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 396 rows and 142 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(San_comuna2)
}
\description{
A dataset containing information about whole survey and geographic features of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{

}
\examples{
 data(San_comuna2)
 COMUNA <- San_comuna2$COMUNA
}
\keyword{datasets}
