% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{package}
\name{Santiago-package}
\alias{Santiago-package}
\title{A data package with the results of a travel and well-being survey conducted in Santiago in 2016}
\description{
A data package based on a paper-based survey conducted face-to-face in Santiago in 2016.
The goal of Santiago is to contribute to psychological impacts on travelers by examining the
feelings of stress by users of active and motorized modes of transportation.Furthermore, it
also investigates the importance that travelers attach to their feelings of stress.This package
contains different datasets namely as Individual characteristics (IC), Health (H), Feelings and
emotions (FE), Reason and planning own decision (RPD), Social interaction (SI), Nature and
sustainability (NS), ITC and shifting (IS), Build environment(BE), Traveling to work (TW).
}
\references{
\url{https://github.com/NiloofarNL/Santiago}
}
\author{
Niloofar Nalaee, School of Earth, Environment and Society, McMaster University \email{nalaeen@mcmaster.ca}
}
