% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_BE}
\alias{Santiago_BE}
\title{A data frame with information about the build environment near the place of residence of respondents to the survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 22 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_BE)
}
\description{
A dataset containing information about Built environment near the users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 22 variables:
\describe{
\item{r7AA_AUTOSPACE}{How do respondents evaluate the following aspects near their home? Space for cars}
\item{r7AB_PARKING_NUMB}{How do respondents evaluate the following aspects near their home? Amount of parking spaces}
\item{r7AC_QHIWAY}{How do respondents evaluate the following aspects near their home? Quality of highways}
\item{r7AD_PEDESTRN}{How do respondents evaluate the following aspects near their home? Space for pedestrians}
\item{r7AE_QSIDEWA}{How do respondents evaluate the following aspects near their home? Quality of sidewalks}
\item{r7AF_CLEAN_STOP}{How do respondents evaluate the following aspects near their home? Cleaningness of bus stops}
\item{r7AG_SEAT}{How do respondents evaluate the following aspects near their home? Sitting areas in bus stops}
\item{r7AH_CLIMA}{How do respondents evaluate the following aspects near their home? Weather protection in bus stops}
\item{r7AI_CICLEWA_NUMB}{How do respondents evaluate the following aspects near their home? Amount of cycleways}
\item{r7AJ_CICLEWA_Q}{How do respondents evaluate the following aspects near their home? Quality of cycleways}
\item{r7AK_BICSHARE}{How do respondents evaluate the following aspects near their home? Bike sharing schemes}
\item{r7BA_IMPAUTOSPACE}{How do respondents rate the level of importance of: Space for cars}
\item{r7BB_IMPPARKING_NUMB}{How do respondents rate the leve of importance of: Amount of parking spaces}
\item{r7BC_IMPQHIWAY}{How do respondents rate the leve of importance of: Quality of highways}
\item{r7BD_IMPPEDESTRN}{How do respondents rate the leve of importance of: Space for pedestrians}
\item{r7BE_IMPQSIDEWA}{How do respondents rate the leve of importance of: Quality of sidewalks}
\item{r7BF_IMPCLEAN_STOP}{How do respondents rate the leve of importance of: Cleaningness of bus stops}
\item{r7BG_IMPSEAT}{How do respondents rate the leve of importance of: Sitting areas in bus stops}
\item{r7BH_IMPCLIMA}{How do respondents rate the leve of importance of: Weather protection in bus stops}
\item{r7BI_IMPCICLEWA_NUMB}{How do respondents rate the leve of importance of: Amount of cycleways}
\item{r7BJ_IMPCICLEWA_Q}{How do respondents rate the leve of importance of: Quality of cycleways}
\item{r7BK_IMPBICSHARE}{How do respondents rate the leve of importance of: Bike sharing schemes}
}
}
\examples{
 data(Santiago_BE)
 r7AA_AUTOSPACE <- Santiago_BE$r7AA_AUTOSPACE
}
\keyword{2016.}
\keyword{Santiago}
\keyword{built}
\keyword{environment}
\keyword{in}
\keyword{near}
\keyword{travelers}
