% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_FE}
\alias{Santiago_FE}
\title{A data frame with information about the feelings and emotions of respondents to the survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 23 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_FE)
}
\description{
A dataset containing information about Feelings and emotions of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 22 variables:
\describe{
\item{ID}{Unique identifier of respondent}
\item{FE_Freedom}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Freedom; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Unsafety}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Unsafety; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Functionality}{indicatation of of the modes of transport that respondents relate to the following feelings and concepts: Functionality; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Enjoyment}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Enjoyment; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Low_cost}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Low cost; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Poverty}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Poverty; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Safety_mode}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Safety; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Waste_time}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Waste of time; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Unpunctuality}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Unpunctuality; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Congestion}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Congestion; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Efficiency}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Efficiency; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Luxury}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Luxury; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Envrnmt_care}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Environmental care; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Health}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Health; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Social_Interaction}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Social Interaction; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Uncomfort}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Uncomfort; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Happiness}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Happiness; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Status}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Status; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{FE_Daily_enjoy}{How much do respondents usually enjoy traveling to their daily activities?; this variable is an ordered factor with five levels, "VHIGH", "HIGH", "MEDIUM", "LOW", "VLOW"}
\item{FE_Imp_quality}{What level of importance do respondents attribute to the quality of their trips?; this variable is an ordered factor with five levels, "VHIGH", "HIGH", "MEDIUM", "LOW", "VLOW"}
\item{FE_Affect}{What does affect the most to their travel experience?; this variable is a factor with eight levels, "COMFORT ABSENCE", "SAFETY ABSENCE", "POOR QUALITY OF STREETS", "CROWDNESS OF PASSENGERS", "DISABILITY OR REDUCED MOBILITY", "TRAVEL TIME", "ALL OF THEM", "OTHER"}
\item{FE_Facility}{What does help respondents the most to their travel experience?; this variable is a factor with eight levels, "COMFORT IMPROVEMENT", "SAFETY IMPROVEMENT", "BETTER QULALITY ON STREETS", "LESS CROWDNESS OF PASSENGERS", "IMPROVEMENT ON ACCESIBILITY", "REDUCTION OF TIME TRAVEL", "ALL OF THEM", "OTHER"}
}
}
\examples{
 data(Santiago_FE)
 r2AA_FREEDOM <- Santiago_FE$r2AA_FREEDOM
}
\keyword{2016.}
\keyword{Santiago}
\keyword{and}
\keyword{emotions}
\keyword{feelings}
\keyword{in}
\keyword{of}
\keyword{travelers}
