% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_IC}
\alias{Santiago_IC}
\title{A data frame with socio-demographic information of respondents to the survey}
\format{
A data frame with 451 rows and 22 variables:
\describe{
\item{ID}{Unique identifier of respondent}
\item{IC_Residence}{Commune of residence}
\item{IC_Macro_res}{Macro area of residence}
\item{IC_Home_St}{Intersections close to your home}
\item{IC_Home_St2}{Intersections close to your home}
\item{IC_Gender}{Sex of respondents in Santiago in 2016; this variable is a factor with two levels, "FEMALE" and "MALE"}
\item{IC_Age}{Age of respondents in Santiago in 2016; this variable is an ordered factor with Six levels, "Less18", "18to24", "25to34", "35to54", "55to64", "65more"}
\item{IC_Weight}{Weight of respondents in kg}
\item{IC_Height}{Height of respondents in cm}
\item{IC_Education}{Level of education of respondents in Santiago in 2016; this variable is an ordered factor with five levels, "ELEMENTARY", "SECUNDARY", "PROFESS_TECH", "COLLEGE", "POSTGRAD"}
\item{IC_Occupation}{Occupation of respondents in Santiago in 2016; this variable is a factor with nine levels, "FULL_TIME", "PART_TIME", "SELF_EMPLOYED", "UNEMPLOYED", "HOME_TAKER", "STUDENT", "STUDENT_WORK", "RETIRED", "OTHER"}
\item{IC_Work_St}{Commune where you perform your work/studies/regular trips}
\item{IC_Work_St2}{Commune where you perform your second set of work/studies/regular trips}
\item{IC_Income}{Level of income of respondents in Santiago in 2016; this variable is an ordered factor with six levels, "Less423", "423to639", "639to977", "977to1550", "1550to2380", "2380more" and the levels are pesos}
\item{IC_Adults}{Number of adults living at home}
\item{IC_Children}{Number of children iving at home}
\item{IC_Drv_license}{Having driving licence; this variable is a factor with two levels, "YES", "NO"}
\item{IC_Disabil}{Being disable or not; this variable is a factor with two levels, "YES", "NO"}
\item{IC_Dis_type}{Type of disability}
\item{IC_Main_mode}{Main transport mode used by respondents in Santiago in 2016; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{IC_Sec_mode}{Secondary transport mode used by respondents in Santiago in 2016; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{IC_Other_mode}{Other transport mode to add used by respondents in Santiago in 2016; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
}
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_IC)
}
\description{
A dataset containing socio-demographic information of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\examples{
 data(Santiago_IC)
 ID <- Santiago_IC$ID
}
\keyword{2016.}
\keyword{Santiago}
\keyword{in}
\keyword{information}
\keyword{of}
\keyword{socio-demographic}
\keyword{travelers}
