% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_NS}
\alias{Santiago_NS}
\title{A data frame with information about the nature and sustainability from the viewpoint of respondents to the survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 10 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_NS)
}
\description{
A dataset containing information about Nature and sustainability from the viewpoint of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 10 variables:
\describe{
\item{r5A_CHANGE}{In a climatic event like heavy rain or flood, do respondents change their main mode of transport?}
\item{r5B_CHANGE_MODE}{If respondents answered yes (change mode of transport in a weather event), indication of the mode(s) is required}
\item{r5C_SUST}{How do respondents assess their access to the currently available sustainable modes of transport? (eg hybrid buses, electric cars, public bicycles)}
\item{r5D_IMP_SUST}{How important is it for respondents to have access to sustainable modes of transport?}
\item{r5E_PAYMENT}{Would respondents be willing to spend more on transportation to gain access to more sustainable modes?}
\item{r5F_PAYMENTS}{If respondents' previous response was positive, indication of how much their transportation expense would increase:}
\item{r5GA_TREE}{indication of how important it is improving the following aspects in public transport routes: Presence of trees}
\item{r5GB_PARK}{indication of how important it is improving the following aspects in public transport routes: Access to parks}
\item{r5GC_MODE}{indication of how important it is improving the following aspects in public transport routes: Access to sustainable transport modes}
\item{r5GD_MODE}{indication of how important it is improving the following aspects in public transport routes: Broaden supply of sustainable transport modes}
}
}
\examples{
 data(Santiago_NS)
 r5A_CHANGE <- Santiago_NS$r5A_CHANGE
}
\keyword{2016.}
\keyword{Santiago}
\keyword{and}
\keyword{in}
\keyword{nature}
\keyword{of}
\keyword{sustainability}
\keyword{travelers'viewpoint}
