% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_RPD}
\alias{Santiago_RPD}
\title{A data frame with information about the reason and planning decision of respondents to the survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 24 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_RPD)
}
\description{
A dataset containing information about reason and planning decision of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 23 variables:
\describe{
\item{ID}{Unique identifier of respondent}
\item{RPD_Access}{How do respondents assess their access to employment opportunities through public transport?; this variable is an ordered factor with five levels,"POOR", "FAIR", "GOOD", "VERY GOOD", "EXCELLENT"}
\item{RPD_Acc_Employment}{How does their access to public transport allows you to access the employment you need?; this variable is an ordered factor with five levels, "POOR", "FAIR", "GOOD", "VERY GOOD", "EXCELLENT"}
\item{RPD_Family_vis}{How often do respondents perform these activities? Visiting family and friends; this variable is an ordered factor with five levels, "NEVER", "RARELY", "SOMETIMES", "OFTEN", "ALWAYS"}
\item{RPD_Rec_activity}{How often do respondents perform these activities? Recreational activities; this variable is an ordered factor with five levels, "NEVER", "RARELY", "SOMETIMES", "OFTEN", "ALWAYS"}
\item{RPD_Cult_activity}{How often do respondents perform these activities? Cultural activities; this variable is an ordered factor with five levels, "NEVER", "RARELY", "SOMETIMES", "OFTEN", "ALWAYS"}
\item{RPD_Sport}{How often do respondents perform these activities? Sport activities; this variable is an ordered factor with five levels, "NEVER", "RARELY", "SOMETIMES", "OFTEN", "ALWAYS"}
\item{RPD_Groc_shopping}{How often do respondents perform these activities? Grocery/Food shopping; this variable is an ordered factor with five levels, "NEVER", "RARELY", "SOMETIMES", "OFTEN", "ALWAYS"}
\item{RPD_Social_activity}{How often do respondents perform these activities? Social activities; this variable is an ordered factor with five levels, "NEVER", "RARELY", "SOMETIMES", "OFTEN", "ALWAYS"}
\item{RPD_Imp_Family_vis}{What is the importance respondents assign to these aspects? Visiting family and friends; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_Rec_activity}{What is the importance respondents assign to these aspects? Recreational activities; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_Cult_activity}{What is the importance respondents assign to these aspects? Cultural activities; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_Sport}{What is the importance respondents assign to these aspects? Sport activities; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_Groc_shopping}{What is the importance respondents assign to these aspects? Grocery/Food shopping; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_Social_activity}{What is the importance respondents assign to these aspects? Social activities; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_Options}{How important do respondents consider having several options to use different modes of transport?; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Trans_Quality}{How much does respondents' quality of life depend on the access respondents currently have to public transport?; this variable is an ordered factor with five levels, "VLOW", "LOW", "MEDIUM", "HIGH", "VHIGH"}
\item{RPD_Inc_Quality}{How much do respondents think their quality of life would increase if they have better access to public transport?; this variable is an ordered factor with five levels, "VLOW", "LOW", "MEDIUM", "HIGH", "VHIGH"}
\item{RPD_Affordable_mode}{Which transport modes are affordable for respondents and their family?; this variable is a factor with eight levels, "CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{RPD_Unaffordable_mode}{Which transport modes are economically unaffordable?; this variable is a factor with eight levels,"CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
\item{RPD_Imp_commercial}{How important it is to improve the following aspects of the public transport system? Improve access to offices and comerccial areas; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_dispon_mode}{How important it is to improve the following aspects of the public transport system? Improve disponibility of different transport modes; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_imprv_comfort}{How important it is to improve the following aspects of the public transport system? Improve comfort for the use of public transport; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{RPD_Imp_othermodes}{How important it is to improve the following aspects of the public transport system? Incorporate other modes to the fare system; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
}
}
\examples{
 data(Santiago_RPD)
 r3A_ACCESS <- Santiago_RPD$r3A_ACCESS
}
\keyword{2016.}
\keyword{Santiago}
\keyword{and}
\keyword{decision}
\keyword{in}
\keyword{of}
\keyword{planning}
\keyword{reason}
\keyword{travelers}
