% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_SI}
\alias{Santiago_SI}
\title{A data frame with information of the Social interaction of respondents to the survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 4 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_SI)
}
\description{
A dataset containing information about Social interaction of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 4 variables:
\describe{
\item{r4A_INTERACC}{How do respondents rate the level of interaction they have with other people during their usual trips?}
\item{r4B_PERSON}{How important do respondents consider the presence of other people during their usual travels?}
\item{r4C_DISCRIM}{Have respondents ever felt discriminated while traveling in any mode of transport?}
\item{r4D_MODE}{If their answer was yes (respondents have felt discriminated),indication of the mode(s) is required}
}
}
\examples{
 data(Santiago_SI)
 r4A_INTERACC <- Santiago_SI$r4A_INTERACC
}
\keyword{2016.}
\keyword{Santiago}
\keyword{in}
\keyword{interaction}
\keyword{of}
\keyword{social}
\keyword{travelers}
