% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_TW}
\alias{Santiago_TW}
\title{Santiago_TW}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 7 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_TW)
}
\description{
A dataset containing information about Traveling to work users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 7 variables:
\describe{
\item{r8A_ACCESSJOB}{How much do respondents think their access to the transport network has affected their chances of having a better job?}
\item{r8B_JOBOPP}{How do respondents assess the job opportunities available in their commune of residence?}
\item{r8C_ACC_COM}{What level of access to employment are respondents interested in having in their commune of residence?}
\item{r8D_EMPLSATISF}{How satisfied are respondents with their current job?}
\item{r8E_TIMECOMMUT}{How long is your regular commuting?}
\item{r8F_SCHEDULE}{What is the frequent time of their commuting?}
\item{r8G_SPENDING}{Monthly expenditure on transport}
}
}
\examples{
 data(Santiago_TW)
 r8A_ACCESSJOB <- Santiago_TW$r8A_ACCESSJOB
}
\keyword{datasets}
