% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_FE}
\alias{Santiago_FE}
\title{A data frame with information about the feelings and emotions of respondents to the survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 22 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_FE)
}
\description{
A dataset containing information about Feelings and emotions of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 22 variables:
\describe{
\item{r2AA_FREEDOM}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Freedom}
\item{r2AB_UNSAFETY}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Unsafety}
\item{r2AC_FUNCTIONALITY}{indicatation of of the modes of transport that respondents relate to the following feelings and concepts: Functionality}
\item{r2AD_ENJOYMENT}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Enjoyment}
\item{r2AE_LOWCOST}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Low cost}
\item{r2AF_POVERTY}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Poverty}
\item{r2AG_SAFETY}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Safety}
\item{r2AH_WASTE_OF_TIME}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Waste of time}
\item{r2AI_UNPUNCTUALITY}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Unpunctuality}
\item{r2AJ_CONGEST}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Congestion}
\item{r2AK_EFFICIENCY}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Efficiency}
\item{r2AL_LUXURY}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Luxury}
\item{r2AM_ENVIRONMENT}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Environmental care}
\item{r2AN_HEALTH}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Health}
\item{r2AO_INTSOCI}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Social Interaction}
\item{r2AP_UNCOMFT}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Uncomfort}
\item{r2AQ_HAPPINESS}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Happiness}
\item{r2AR_STATUS}{indicatation of the modes of transport that respondents relate to the following feelings and concepts: Status}
\item{r2B_DAILY_ENJOY}{How much do respondents usually enjoy traveling to their daily activities?}
\item{r2C_IMP_QUALITY}{What level of importance do respondents attribute to the quality of their trips?}
\item{r2D_AFFECT}{What does affect the most to their travel experience?}
\item{r2E_FACILIT}{What does help respondents the most to their travel experience?}
}
}
\examples{
 data(Santiago_FE)
 r2AA_FREEDOM <- Santiago_FE$r2AA_FREEDOM
}
\keyword{2016.}
\keyword{Santiago}
\keyword{and}
\keyword{emotions}
\keyword{feelings}
\keyword{in}
\keyword{of}
\keyword{travelers}
