% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_IC}
\alias{Santiago_IC}
\title{A data frame with socio-demographic information of respondents to the survey}
\format{
A data frame with 451 rows and 23 variables:
\describe{
\item{ID}{Unique identifier of respondent}
\item{ORIGIN}{Origin of travel}
\item{r0A_COMMUNE}{Commune of residence}
\item{r0B_MACR_ZONE}{Macro area of residence}
\item{r0B_CALLE_1}{Intersections close to your home}
\item{r0B_CALLE_2}{Intersections close to your home}
\item{r0C_GENDER}{Gender of respondents in Santiago in 2016}
\item{r0D_AGE}{Age of respondents in Santiago in 2016}
\item{r0E_WEIGHT}{Weight of respondents in Santiago in 2016}
\item{r0F_HEIGHT}{Height of respondents in Santiago in 2016}
\item{r0G_EDUCATION}{Level of education of respondents in Santiago in 2016}
\item{r0H_OCCUPATION}{Occupation of respondents in Santiago in 2016}
\item{r0I_COM_TRAB_1}{Communes where you perform your work/studies/regular trips}
\item{r0I_COM_TRAB_2}{Communes where you perform your work/studies/regular trips}
\item{r0J_INCOME}{Level of income of respondents in Santiago in 2016}
\item{r0K_ADULTS}{Number of adults living at home}
\item{r0L_CHILDREN}{Number of children iving at home}
\item{r0N_DRIVING_LICENSE}{Having driving licence}
\item{r0M_DISABILITY}{Being disable or not }
\item{r0O_TYPE_DIS}{Type of disability}
\item{r0P_MODE1}{Main transport mode used by respondents in Santiago in 2016}
\item{r0Q_MODE2}{Secondary transport mode used by respondents in Santiago in 2016}
\item{r0R_MODE3}{Other transport mode to add used by respondents in Santiago in 2016}
}
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_IC)
}
\description{
A dataset containing socio-demographic information of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\examples{
 data(Santiago_IC)
 ID <- Santiago_IC$ID
}
\keyword{2016.}
\keyword{Santiago}
\keyword{in}
\keyword{information}
\keyword{of}
\keyword{socio-demographic}
\keyword{travelers}
