% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_RPD}
\alias{Santiago_RPD}
\title{Santiago_RPD}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 23 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_RPD)
}
\description{
A dataset containing information about reason and planning decision of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 23 variables:
\describe{
\item{r3A_ACCESS}{How do respondents assess their access to employment opportunities through public transport?}
\item{r3B_ACC_EM}{How does their access to public transport allows you to access the employment you need?}
\item{r3CA_FAM}{How often do respondents perform these activities? Visiting family and friends}
\item{r3CB_REC}{How often do respondents perform these activities? Recreational activities}
\item{r3CC_CUL}{How often do respondents perform these activities? Cultural activities}
\item{r3CD_SPO}{How often do respondents perform these activities? Sport activities}
\item{r3CE_GROC}{How often do respondents perform these activities? Grocery/Food shopping}
\item{r3CF_SOC}{How often do respondents perform these activities? Social activities}
\item{r3DA_FAM}{NEEDS DESCRIPTION}
\item{r3DB_REC}{NEEDS DESCRIPTION}
\item{r3DC_CUL}{NEEDS DESCRIPTION}
\item{r3DD_SPO}{NEEDS DESCRIPTION}
\item{r3DE_GROC}{NEEDS DESCRIPTION}
\item{r3DF_SOC}{NEEDS DESCRIPTION}
\item{r3E_OPTIONS}{How important do respondents consider having several options to use different modes of transport?}
\item{r3F_ACCESS_DEPENDENCY}{How much does respondents' quality of life depend on the access respondents currently have to public transport?}
\item{r3G_QUALITY_INCRS}{How much do respondents think their quality of life would increase if they have better access to public transport?}
\item{r3H_ECON}{Which transport modes are affordable for respondents and their family?}
\item{r3I_NOECON}{Which transport modes are economically unaffordable?}
\item{r3JA_OFIC}{How important it is to improve the following aspects of the public transport system? Improve access to offices and comerccial areas}
\item{r3JB_MODES}{How important it is to improve the following aspects of the public transport system? Improve disponibility of different transport modes}
\item{r3JC_COMFORT}{How important it is to improve the following aspects of the public transport system? Improve comfort for the use of public transport}
\item{r3JD_OTHERS}{How important it is to improve the following aspects of the public transport system? Incorporate other modes to the fare system}
}
}
\examples{
 data(Santiago_RPD)
 r3A_ACCESS <- Santiago_RPD$r3A_ACCESS
}
\keyword{2016.}
\keyword{Santiago}
\keyword{and}
\keyword{decision}
\keyword{in}
\keyword{of}
\keyword{planning}
\keyword{reason}
\keyword{travelers}
