% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSantiago-package.R
\docType{data}
\name{Santiago_SI}
\alias{Santiago_SI}
\title{A data frame with information of the Social interaction of respondents to the survey}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 451 rows and 4 columns.
}
\source{
A face to face survey considering pre-census 2012
}
\usage{
data(Santiago_SI)
}
\description{
A dataset containing information about Social interaction of users of active and motorized modes of transportation in
Santiago sourced from 2016 survey.
}
\details{
@format A data frame with 451 rows and 4 variables:
\describe{
\item{ID}{Unique identifier of respondent}
\item{r4A_INTERACC}{How do respondents rate the level of interaction they have with other people during their usual trips?; this variable is an ordered factor with five levels, "POOR", "FAIR", "GOOD", "VERY GOOD", "EXCELLENT"}
\item{r4B_PERSON}{How important do respondents consider the presence of other people during their usual travels?; this variable is an ordered factor with five levels, "NOT IMPORTANT", "SLIGHTLY IMPORTANT", "MODERATELY IMPORTANT", "IMPORTANT", "VERY IMPORTANT"}
\item{r4C_DISCRIM}{Have respondents ever felt discriminated while traveling in any mode of transport?; this variable is an out of order factor with two levels, "YES", "NO"}
\item{r4D_MODE}{If their answer was yes (respondents have felt discriminated),indication of the mode(s) is required; this variable is an out of order factor with eight levels, "CAR", "TAXI", "COLECTIVO", "MOTO", "METRO", "BUS","BICYCLE", "WALK"}
}
}
\examples{
 data(Santiago_SI)
 r4A_INTERACC <- Santiago_SI$r4A_INTERACC
}
\keyword{2016.}
\keyword{Santiago}
\keyword{in}
\keyword{interaction}
\keyword{of}
\keyword{social}
\keyword{travelers}
